+++ ClinicalLab/ClinicalLab-Infrastructure/src/main/java/infrastructure/persistence/jpa/repository/ExamRequestRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import infrastructure.persistence.jpa.ExamRequestJPA;
 
-public interface ExamRequestRepository extends JpaRepository<ExamRequestJPA, Long>{
+public interface ExamRequestRepository extends JpaRepository<ExamRequestJPA, Integer>{
 
 }
